Project/backend/requirements.txt
absl-py==0.10.0
airsim==1.2.8
antiorm==1.2.1
asgiref==3.2.10
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.3
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.0
bootstrap4==0.1.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
cycler==0.10.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-bootstrap==0.2.4
django-cors-headers==3.5.0
django-extensions==2.2.9
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
dlib==19.21.0
docopt==0.6.2
drf-yasg==1.17.1
entrypoints==0.3
Flask==1.1.2
gast==0.2.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grip==4.5.2
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
inflection==0.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.1.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.3
pandas==1.1.2
pandocfilters==1.4.2
parso==0.5.2
path-and-address==2.0.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.15.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
tb-nightly==2.3.0a20200401
tensorboard==2.1.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==4.5.3
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.2.0
